+++ R/render_xlsx.R
@@ -1,9 +1,3 @@
-require_xlsx <- function() {
-  if (!requireNamespace("xlsx", quietly = TRUE)) {
-    stop("Please install the xlsx package in order to export to excel")
-  }
-}
-
 #' Writes the table to an Excel workbook
 #'
 #' @param x A condformat_tbl object
@@ -16,37 +10,27 @@ require_xlsx <- function() {
 condformat2excel <- function(x, filename, sheet_name = "Sheet1",
                              overwrite_wb = FALSE,
                              overwrite_sheet = TRUE) {
-  require_xlsx()
 
   if (!grepl(pattern = '\\.xlsx$', filename)) { # endsWith(filename, ".xlsx")
     filename <- paste0(filename, ".xlsx")
   }
 
   if (file.exists(filename) && identical(overwrite_wb, FALSE)) {
-    wb <- xlsx::loadWorkbook(filename)
+    wb <- openxlsx::loadWorkbook(filename)
   } else {
-    wb <- xlsx::createWorkbook(type = "xlsx")
+    wb <- openxlsx::createWorkbook(creator = "")
   }
 
-  # getSheets cat's a message I don't want nor care about.
-  noSheets <- wb$getNumberOfSheets()
-  if (noSheets > 0) {
-    sheet_list <- xlsx::getSheets(wb)
-  } else {
-    sheet_list <- list()
-  }
-  if (sheet_name %in% names(sheet_list)) {
+  if (sheet_name %in% names(wb)) {
     if (overwrite_sheet) {
-      xlsx::removeSheet(wb = wb, sheetName = sheet_name)
-      sheet <- xlsx::createSheet(wb, sheetName = sheet_name)
-    } else {
-      sheet <- sheet_list[[sheet_name]]
+      openxlsx::removeWorksheet(wb = wb, sheet = sheet_name)
+      openxlsx::addWorksheet(wb, sheetName = sheet_name)
     }
   } else {
-    sheet <- xlsx::createSheet(wb, sheetName = sheet_name)
+    openxlsx::addWorksheet(wb, sheetName = sheet_name)
   }
-  condformat2excelsheet(x, sheet)
-  xlsx::saveWorkbook(wb, file = filename)
+  condformat2excelsheet(x, wb, sheet_name)
+  openxlsx::saveWorkbook(wb, file = filename)
   return(invisible(x))
 }
 
@@ -57,24 +41,20 @@ condformat2excel <- function(x, filename, sheet_name = "Sheet1",
 # @examples
 # \dontrun{
 # x <- condformat(iris[1:5,])
-# library(xlsx)
-# wb <- xlsx::createWorkbook(type = "xlsx")
-# sheet <- xlsx::createSheet(wb, sheetName = "Sheet1")
-# condformat2excelsheet(x, sheet)
-# xlsx::saveWorkbook(wb, file = "iris.xlsx")
+# library(openxlsx)
+# wb <- openxlsx::createWorkbook(creator = "")
+# openxlsx::addWorksheet(wb, sheetName = "sheet name")
+# condformat2excelsheet(x, wb, "sheet name")
+# openxlsx::saveWorkbook(wb, file = "iris.xlsx")
 # }
-condformat2excelsheet <- function(x, sheet) {
-  require_xlsx()
-  if (!"jobjRef" %in% class(sheet)) {
-    stop("sheet must be an jobjRef object, as the one returned with xls::createSheet()")
-  }
+condformat2excelsheet <- function(x, wb, sheet_name) {
   xv_cf <- get_xview_and_cf_fields(x)
   xview <- xv_cf[["xview"]]
   cf_fields <- xv_cf[["cf_fields"]]
 
   css_fields <- render_cf_fields_to_css_fields(cf_fields, xview)
-  xlsx::addDataFrame(x = as.data.frame(xview),
-                     sheet = sheet, row.names = FALSE, col.names = TRUE)
+  openxlsx::writeData(wb, sheet_name, as.data.frame(xview),
+                      rowNames = FALSE, colNames = TRUE)
   for (css_key in names(css_fields)) {
     if (css_key == "background-color") {
       for (i in seq_len(nrow(xview))) {
@@ -83,12 +63,8 @@ condformat2excelsheet <- function(x, sheet) {
                                      NA,
                                      css_fields[["background-color"]][i,j])
           if (!is.na(background_color)) {
-            cb <- xlsx::CellBlock(sheet, startRow = i + 1, startColumn = j,
-                                  noRows = 1, noColumns = 1, create = FALSE)
-            fill <- xlsx::Fill(backgroundColor = background_color, foregroundColor = background_color)
-            xlsx::CB.setFill(cellBlock = cb,
-                             fill = fill,
-                             rowIndex = 1, colIndex = 1)
+            sty <- openxlsx::createStyle(fgFill = background_color)
+            openxlsx::addStyle(wb, sheet_name, style = sty, rows = i + 1, cols = j)
           }
         }
       }