+++ R/rule_css.R
@@ -7,9 +7,12 @@
 #' @param na.value CSS property value to be used in missing values (e.g. `"grey"`)
 #' @examples
 #' data(iris)
-#' condformat(iris[c(1:5, 51:55, 101:105),]) %>%
+#' cf <- condformat(iris[c(1:5, 51:55, 101:105),]) %>%
 #'   rule_css(Species, expression = ifelse(Species == "setosa", "red", "darkgreen"),
 #'            css_field = "color")
+#' \dontrun{
+#' print(cf)
+#' }
 #' @export
 rule_css <- function(x, columns, expression,
                      css_field,