+++ R/rule_fill_gradient2.R
@@ -20,13 +20,19 @@
 #' @return The condformat_tbl object, with the added formatting information
 #' @examples
 #' data(iris)
-#' condformat(iris[c(1:5, 70:75, 120:125), ]) %>%
+#' cf <- condformat(iris[c(1:5, 70:75, 120:125), ]) %>%
 #'  rule_fill_gradient2(Sepal.Length) %>%
 #'  rule_fill_gradient2(Species, expression=Sepal.Length - Sepal.Width)
+#' \dontrun{
+#' print(cf)
+#' }
 #'
-#' condformat(iris[c(1:5, 70:75, 120:125), ]) %>%
+#' cf <- condformat(iris[c(1:5, 70:75, 120:125), ]) %>%
 #'   rule_fill_gradient2("Petal.Length") %>%
 #'   rule_fill_gradient2(starts_with("Sepal"), expression=Sepal.Length - Sepal.Width)
+#' \dontrun{
+#' print(cf)
+#' }
 #'
 #' @importFrom scales rescale_mid
 #' @export