+++ R/rule_text_bold.R
@@ -6,8 +6,11 @@
 #' @param na.bold If `TRUE`, make missing values bold.
 #' @examples
 #' data(iris)
-#' condformat(iris[c(1:5, 51:55, 101:105),]) %>%
+#' cf <- condformat(iris[c(1:5, 51:55, 101:105),]) %>%
 #'   rule_text_bold(Species, expression = Species == "setosa")
+#' \dontrun{
+#' print(cf)
+#' }
 #' @export
 rule_text_bold <- function(x, columns, expression,
                            na.bold = FALSE,