+++ R/rule_text_color.R
@@ -6,8 +6,11 @@
 #' @param na.color Color for missing values
 #' @examples
 #' data(iris)
-#' condformat(iris[c(1:5, 51:55, 101:105),]) %>%
+#' cf <- condformat(iris[c(1:5, 51:55, 101:105),]) %>%
 #'   rule_text_color(Species, expression = ifelse(Species == "setosa", "blue", ""))
+#' \dontrun{
+#' print(cf)
+#' }
 #' @export
 rule_text_color <- function(x, columns, expression,
                             na.color = "",